+++ app/robots.ts
@@ -7,6 +7,6 @@ export default function robots(): MetadataRoute.Robots {
       allow: "/",
       disallow: "/api/",
     },
-    sitemap: "https://chatdevai.com/sitemap.xml",
+    sitemap: "https://agentdevai.com/sitemap.xml",
   }
 }
