+++ components/structured-data/organization-schema.tsx
@@ -5,15 +5,15 @@ export function OrganizationSchema() {
     name: "AgentDevAI",
     description:
       "Professional AI agent development with LangChain, LangGraph, and FastAPI. Build scalable WhatsApp automation and multi-channel integration.",
-    url: "https://chatdevai.com",
-    logo: "https://chatdevai.com/chatdevai-logo.png",
+    url: "https://agentdevai.com",
+    logo: "https://agentdevai.com/agentdevai-logo.png",
     foundingDate: "2023",
     contactPoint: {
       "@type": "ContactPoint",
       contactType: "Customer Service",
-      email: "contact@chatdevai.com",
+      email: "contact@agentdevai.com",
     },
-    sameAs: ["https://twitter.com/chatdevai", "https://linkedin.com/company/chatdevai", "https://github.com/chatdevai"],
+    sameAs: ["https://twitter.com/agentdevai", "https://linkedin.com/company/agentdevai", "https://github.com/agentdevai"],
     address: {
       "@type": "PostalAddress",
       addressCountry: "US",
